+++ services/userService.js
@@ -1,7 +1,6 @@
 import { userRepository } from "../repositories/userRepository.js";
 
 class UserService {
-  // TODO: Implement methods to work with user
 
   update(id, dataToUpdate) {
     const updatedUser = userRepository.update(id, dataToUpdate);
@@ -19,7 +18,12 @@ class UserService {
     return user;
   }
   delete(id) {
-    return userRepository.delete(id)
+    const user = this.getOne({id})
+    if(!!user) {
+      return userRepository.delete(id)
+    }
+    return null
+
   }
   getAll() {
     const users = userRepository.getAll()
@@ -42,6 +46,29 @@ class UserService {
       return null;
     }
     return item;
+  }
+
+  isEmailValid = (email) => {
+    return /^\w+([\.-]?\w+)*@\w+([\.-]?\w+)*(\.\w{2,3})+$/.test(email)
+  }
+
+  isPhoneValid = (phone) => {
+    return phone.startsWith("+380") && phone.length == 13
+  }
+
+  getRedundantKeys = (keys, userKeys) => {
+    const redundantKeys = userKeys.filter(item => !keys.includes(item))
+    return redundantKeys.length > 0
+  }
+
+  isUserWithSameEmail(email) {
+    const allUsers  = this.getAll();
+    return allUsers.find(item => item.email === email);
+  }
+
+  isUserWithSamePhone(phone) {
+    const allUsers  = this.getAll();
+    return allUsers.find(item => item.phoneNumber === phone);
   }
 }
